Fixes warning.

+++ src/generators/cpp.ml
@@ -212,7 +212,7 @@ let rec print_exp state (prec : operator option) (e : exp) =
       {%pla|(<#se1#> >> <#n#i>)|})
   | ECall { path = "size"; args = [ e1 ] } ->
     let e1 = print_exp state prec e1 in
-    {%pla|<#e1#>.size()|}
+    {%pla|static_cast<int32_t>(<#e1#>.size())|}
   | ECall { path = "length"; args = [ e1 ] } ->
     let e1 = print_exp state prec e1 in
     {%pla|static_cast<int32_t>(<#e1#>.size())|}
